requirements.txt
feedparser==6.0.12
Pillow==12.0.0
pygame==2.6.1
python-dotenv==1.2.1
Requests==2.32.5
